Allow dynamic update of build settings by pipelines.
This change also adds validation of configuration parameters used by install,
and a hook that calls OpenDevice on the intended platform. While we will fail
if the platform does not exist, we only issue a warning if we are unable to
open the device. This change required minor plumbing for the way runsc flags
are registered in order to allow for a flag.FlagSet in core API methods.

PiperOrigin-RevId: 411156716